// frontend/src/api/generated/model/addProtocolItemViewModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { CountUnit } from './countUnit';
import type { PlantOrgan } from './plantOrgan';
import type { ReferenceLevelViewModel } from './referenceLevelViewModel';

export interface AddProtocolItemViewModel {
  targetId: string;
  organ: null | PlantOrgan;
  unit: CountUnit;
  photosRequested: number;
  instructions: string;
  referenceLevel: null | ReferenceLevelViewModel;
}